+++ src/index.js
@@ -1,10 +1,11 @@
 import React, { useState } from 'react'
-import { format as fnsFormat } from 'date-fns'
+import { format as fnsFormat, compareAsc } from 'date-fns'
 import { ThemeProvider } from 'styled-components'
 
 import { defaultTheme } from './components/defaultTheme'
 import { formatDate } from './helpers/formatDate'
 import { mergeDeep } from './helpers/mergeDeep'
+import { getDefaultDate } from './helpers/calendar/defaultDate'
 import SelectDate from './components/SelectDate'
 import Calendar from './components/Calendar'
 
@@ -19,7 +20,9 @@ export const DatePicker = ({
   inputId,
   initialDate
 }) => {
-  const defaultDate = dateRange ? new Date(dateRange.max) : new Date()
+  
+  let defaultDate = getDefaultDate({dateRange})
+
   const [showCalendar, setShowCalendar] = useState(false)
   const [date, setDate] = useState(fnsFormat(new Date(defaultDate), format))
   const [hasSelected, setHasSelected] = useState(false)