+++ src/components/SelectDate.js
@@ -6,10 +6,12 @@ import { format as fnsFormat } from 'date-fns'
 import * as S from './styles'
 import { isValidDate } from '../helpers/isValidDate'
 import { dateFromString } from '../helpers/formatDate'
+import { createUid } from '../helpers/uidMap'
 
 const KEYCODE = {
   ENTER: 13,
-  SPACE: 32
+  SPACE: 32,
+  BACKSPACE: 8
 }
 
 const SelectDate = ({
@@ -24,9 +26,12 @@ const SelectDate = ({
   dateRange,
   onInvalidDate,
   inputId,
-  initialDate
+  initialDate,
+  ariaLabelFormatHint
 }) => {
-  const initDate = initialDate ? initialDate : hasSelected ? date : placeholder
+  const uid = createUid()
+  const defaultFormatHintAriaLabel = `format as ${formatHint}`
+  const initDate = initialDate || (hasSelected ? date : placeholder)
   const [dateInput, setDateInput] = useState(initDate)
 
   useEffect(() => {
@@ -36,7 +41,6 @@ const SelectDate = ({
           value: initialDate
         }
       })
-      //setDateInput(initialDate)
     }
   }, [initialDate])
 
@@ -55,39 +59,59 @@ const SelectDate = ({
     return str
   }
 
+  const onInputKeyDown = e => {
+    if (e.keyCode === KEYCODE.BACKSPACE) {
+      const lastChar = dateInput.charAt(dateInput.length - 1)
+      if (lastChar === '/') {
+        setDateInput(dateInput.substr(0, dateInput.length - 1))
+      }
+    }
+  }
+
+  const onInputKeyUp = e => {
+    if (dateInput.length === 10) {
+      submitInputDate({ e })
+    }
+  }
+
   const onInputKeyPress = e => {
     const charCode = e.charCode
-
     if (charCode === KEYCODE.ENTER) {
-      if (dateInput.length < 10) {
-        onInvalidDate({
-          message: 'date input invalid',
-          details: {
-            dateRange,
-            dateInput
-          }
-        })
-        e.preventDefault()
-        return false
-      }
-      const isValid = isValidDate({ dateRange, str: dateInput })
-
-      if (!isValid) {
-        onInvalidDate({
-          message: 'date outside of range',
-          details: {
-            dateRange,
-            dateInput
-          }
-        })
-        e.preventDefault()
-        return false
-      }
-
-      // set selected date
-      const date = dateFromString({ date: dateInput })
-      handleSelectDate(fnsFormat(date, format))
+      submitInputDate({ e })
     }
+  }
+
+  const submitInputDate = ({ e }) => {
+    if (dateInput.length < 10) {
+      onInvalidDate({
+        message: 'date input invalid',
+        details: {
+          dateRange,
+          dateInput
+        }
+      })
+      e.preventDefault()
+      return false
+    }
+    const isValid = isValidDate({ dateRange, str: dateInput })
+
+    if (!isValid) {
+      onInvalidDate({
+        message: 'date outside of range',
+        details: {
+          dateRange,
+          dateInput
+        }
+      })
+      e.preventDefault()
+      return false
+    }
+    // clear errors
+    onInvalidDate('')
+
+    // set selected date
+    const date = dateFromString({ date: dateInput })
+    handleSelectDate(fnsFormat(date, format))
   }
 
   const onIconKeyPress = e => {
@@ -154,21 +178,22 @@ const SelectDate = ({
       aria-label='Datepicker'
     >
       <S.CalendarUl>
-        <S.InputLi onKeyPress={onInputKeyPress}>
+        <S.InputLi onKeyPress={onInputKeyPress} onKeyUp={onInputKeyUp}>
           <S.DateInput
             id={inputId || 'a11y-date-input'}
             tabIndex='0'
             aria-label='Date Input'
-            aria-describedby='formatHint'
+            aria-describedby={`formatHint-${uid}`}
             onClick={onInputFocus}
             onFocus={onInputFocus}
             onBlur={onInputBlur}
             value={dateInput}
             onChange={(e) => onInputChange(e)}
+            onKeyDown={(e) => onInputKeyDown(e)}
           />
           <S.InputHint
-            aria-label={`format as ${formatHint}`}
-            id='formatHint'
+            aria-label={ariaLabelFormatHint || defaultFormatHintAriaLabel}
+            id={`formatHint-${uid}`}
           >
             {formatHint}
           </S.InputHint>