+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 - parameterized aria-label for formatHint as `ariaLabelFormatHint`
 - if cur date > max date, cal ui will start on max date not current
 - if cur date < max date, cal ui will start on cur date
+- on entering a full date will submit (enter key) via date input
 
 ### Fixed
 - updating initialDate after an API call now triggers correctly
-- fixed bug with having multiple datepickers on page s
+- fixed bug with having multiple datepickers on pages
+- backspace fixed when going back from added slashes i.e. `02/02/` <-- will now give `02/0`
 
 ## [0.0.6-rc.10] - 2021-10-06
 ### Lost