+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 ## [0.0.7-rc.0] - 2021-10-21
+### Added
+- parameterized aria-label for formatHint as `ariaLabelFormatHint`
+- if cur date > max date, cal ui will start on max date not current
+- if cur date < max date, cal ui will start on cur date
+- on entering a full date will submit (enter key) via date input and clear errors
+
 ### Fixed
 - updating initialDate after an API call now triggers correctly
-- fixed bug with having multiple datepickers on page s
+- fixed bug with having multiple datepickers on pages
+- backspace fixed when going back from added slashes i.e. `02/02/` <-- will now give `02/0`
+
+### Changed
+- months for aria-labels are now full i.e. January vs Jan
 
 ## [0.0.6-rc.10] - 2021-10-06
 ### Lost